Migrate a bunch of old codegen tests.
