+++ tests/c/awkward_unmappable.c
@@ -1,5 +1,3 @@
-// ## Fails with: Assertion `!CallStack.curMappableFrame()' failed
-// ignore-if: true
 // Run-time:
 //   env-var: YKD_SERIALISE_COMPILATION=1
 //   status: success
