+++ tests/c/choice.c
@@ -1,39 +1,36 @@
+// Compiler:
 // Run-time:
+//   env-var: YK_LOG=255
 //   env-var: YKD_LOG_IR=-:jit-pre-opt
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
 //   stderr:
-//     jitstate: start-tracing
-//     4:21
-//     jitstate: stop-tracing
+//     ...
+//     yk-jit-event: start-tracing
+//     3: 47
+//     yk-jit-event: stop-tracing
 //     --- Begin jit-pre-opt ---
 //     ...
 //     --- End jit-pre-opt ---
-//     3:21
-//     jitstate: enter-jit-code
-//     2:21
-//     1:21
-//     jitstate: deoptimise
+//     2: 47
+//     yk-jit-event: enter-jit-code
+//     1: 47
+//     yk-jit-event: deoptimise
 //     ...
 
-// Check that recursive function calls are not unrolled.
+// Check that tracing a cascading "if...else if...else" works.
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <yk.h>
 #include <yk_testing.h>
 
-__attribute__((noinline)) int fib(int num) {
-  if (num == 0)
-    return 0;
-  if (num == 1)
-    return 1;
-  if (num == 2)
-    return 1;
-  int a = fib(num - 2);
-  int b = fib(num - 1);
-  int c = a + b;
-  return c;
+__attribute__((noinline)) int f(int x) {
+  if (x == 0)
+    return 30;
+  else if (x == 1)
+    return 47;
+  else
+    return 52;
 }
 
 int main(int argc, char **argv) {
@@ -41,12 +38,12 @@ int main(int argc, char **argv) {
   yk_mt_hot_threshold_set(mt, 0);
   YkLocation loc = yk_location_new();
 
-  int i = 4;
+  int i = 3, x = 1;
   NOOPT_VAL(i);
   while (i > 0) {
     yk_mt_control_point(mt, &loc);
-    NOOPT_VAL(argc);
-    fprintf(stderr, "%d:%d\n", i, fib(argc * 8));
+    NOOPT_VAL(x);
+    fprintf(stderr, "%d: %d\n", i, f(x));
     i--;
   }
