+++ tests/c/intrinsics.c
@@ -1,47 +1,49 @@
 // Run-time:
 //   env-var: YKD_LOG_IR=-:jit-pre-opt
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stderr:
-//     jitstate: start-tracing
-//     3:5
-//     jitstate: stop-tracing
+//     yk-jit-event: start-tracing
+//     yk-jit-event: stop-tracing
 //     --- Begin jit-pre-opt ---
 //     ...
 //     --- End jit-pre-opt ---
-//     2:5
-//     jitstate: enter-jit-code
-//     1:5
-//     jitstate: deoptimise
+//     yk-jit-event: enter-jit-code
 //     ...
+//     yk-jit-event: deoptimise
+//     ...
+//   stdout:
+//     998
 
-// Check that using an argument (of a non-main() function) in a trace works.
+// Check that inlined intrinsics are handled correctly.
 
+#include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <yk.h>
 #include <yk_testing.h>
 
-int f(int x) {
+int main(int argc, char **argv) {
+  int res = 0;
+  int src = 1000;
   YkMT *mt = yk_mt_new(NULL);
   yk_mt_hot_threshold_set(mt, 0);
   YkLocation loc = yk_location_new();
-
   int i = 3;
+  NOOPT_VAL(res);
   NOOPT_VAL(i);
+  NOOPT_VAL(src);
   while (i > 0) {
     yk_mt_control_point(mt, &loc);
-    fprintf(stderr, "%d:%d\n", i, x);
+    memcpy(&res, &src, 4);
+    src--;
     i--;
   }
-
+  NOOPT_VAL(res);
+  printf("%d", res);
   yk_location_drop(loc);
   yk_mt_shutdown(mt);
-  return 0;
-}
 
-int main(int argc, char **argv) {
-  f(5);
   return (EXIT_SUCCESS);
 }
