+++ tests/c/loopy_funcs_not_inlined_by_default.c
@@ -4,17 +4,17 @@
 //   stderr:
 //     --- Begin aot ---
 //     ...
-//     call void @never_inline_into_trace(i32 noundef ...
+//     func main(...
+//     ...
+//     call never_inline_into_trace(%{{9_0}})...
+//     ...
+//     }
 //     ...
 //     --- End aot ---
 //     ...
 //     --- Begin jit-pre-opt ---
 //     ...
-//     define ptr @__yk_compiled_trace_0(ptr %0, ptr %1) {
-//       ...
-//       call void @never_inline_into_trace(i32 noundef ...
-//       ...
-//     }
+//     call @never_inline_into_trace(%{{6}})
 //     ...
 //     --- End jit-pre-opt ---
