+++ tests/c/many_threads_one_loc.c
@@ -1,9 +1,20 @@
-// ## Shadow stack doesn't currently support dynamically sized stack.
+// ## Shadow stack isn't thread safe.
 // ignore-if: true
 // Run-time:
+//   env-var: YKD_LOG_IR=-:aot
+//   env-var: YKD_SERIALISE_COMPILATION=1
 //   env-var: YKD_LOG_JITSTATE=-
 //   stderr:
 //     ...
+//     --- Begin aot ---
+//     ...
+//     define i8* @trace(...
+//       ...
+//       %{{a}} = add i64 %{{b}}, -1...
+//       ...
+//     }
+//     ...
+//     --- End aot ---
 //     jitstate: enter-jit-code
 //     ...
 
@@ -18,10 +29,7 @@
 #include <yk.h>
 #include <yk_testing.h>
 
-#ifdef linux
-#include <sys/sysinfo.h>
-#endif
-
+#define NUM_THREADS 8
 #define ITERS 100000
 
 struct thread_data {
@@ -45,22 +53,15 @@ static void *trace(void *arg) {
 }
 
 int main() {
+  YkLocation loc = yk_location_new();
   YkMT *mt = yk_mt_new(NULL);
   yk_mt_hot_threshold_set(mt, 0);
 
-#ifdef linux
-  int n_thr = get_nprocs();
-#else
-#error unimplemented
-#endif
-
-  pthread_t threads[n_thr];
-  struct thread_data tds[n_thr];
-  YkLocation locs[n_thr];
-  for (int i = 0; i < n_thr; i++) {
-    locs[i] = yk_location_new();
+  pthread_t threads[NUM_THREADS];
+  struct thread_data tds[NUM_THREADS];
+  for (int i = 0; i < NUM_THREADS; i++) {
     tds[i].tnum = i;
-    tds[i].loc = &locs[i];
+    tds[i].loc = &loc;
     tds[i].mt = mt;
 
     if (pthread_create(&threads[i], NULL, trace, &tds[i]) != 0)
@@ -68,13 +69,14 @@ int main() {
   }
 
   void *thread_res = 0;
-  for (int i = 0; i < n_thr; i++) {
+  for (int i = 0; i < NUM_THREADS; i++) {
     if (pthread_join(threads[i], &thread_res) != 0)
       err(EXIT_FAILURE, "pthread_join");
     assert((uintptr_t)thread_res == i);
-    yk_location_drop(locs[i]);
   }
 
+  yk_location_drop(loc);
   yk_mt_shutdown(mt);
+
   return (EXIT_SUCCESS);
 }
